refac: extract validation logic outside component, make ir reusable
